% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGeneProfiles.R
\name{computeGeneProfiles}
\alias{computeGeneProfiles}
\title{Function that computes the gene profiles.}
\usage{
computeGeneProfiles(expressionMatrix, C)
}
\arguments{
\item{expressionMatrix}{The expression matrix organized as (*n* X *m*), *n* is the number of genes and *m* is the number of samples.
You should use the expressionMatrix output of prepareData().}

\item{C}{The C matrix. You should use the Copt output of multivariateCurveResolution().}
}
\value{
The gene profiles data.frame.
}
\description{
Function that computes the gene profiles.
}
\author{
Tobias K. Karakach, Federico Taverna
}
