% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlProjection.R
\name{mlProjection}
\alias{mlProjection}
\title{Function that computes the Maximum Likelihood PCA with equal row correlated errors.}
\usage{
mlProjection(
  expressionMatrix,
  errorCovarianceMatrix,
  numberOfComponents = 15,
  maxIterations = 2000,
  tolerance = 1e-10,
  verbose = TRUE
)
}
\arguments{
\item{expressionMatrix}{The expression matrix organized as (*n* X *m*), *n* is the number of genes and *m* is the number of samples.
You should use the expressionMatrix output of prepareData().}

\item{errorCovarianceMatrix}{The inverse of the measurement errors, whose dimensions are (*n* X *n*).
You should use the errorCovarianceMatrix output of prepareData().}

\item{numberOfComponents}{The number of principal components estimated to be required to reconstruct the expression matrix without loss of generality.}

\item{maxIterations}{The maximum number of iterations to peform during the computation of the Maximum Likelihood. Once the max iteration is reached,
the algorithms stops.}

\item{tolerance}{The termination tolerance for the computation of the Maximum Likelihood. Once the improvements made by Maximum Likelihood reach the
tolerance, the algorithm stops.}

\item{verbose}{Whether to display the information about the computation or not.}
}
\value{
A list consisting of U (principal components), S(eigenvalues) and V (loadings).
The estimated matrix (estimatedMatrix) is given as t(U %*% S %*% t(V)).
}
\description{
Function that computes the Maximum Likelihood PCA with equal row correlated errors.
}
\author{
Tobias K. Karakach, Federico Taverna
}
