% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateCurveResolution.R
\name{multivariateCurveResolution}
\alias{multivariateCurveResolution}
\title{Function to perform multivariate curve resolution by alternating total least squares.}
\usage{
multivariateCurveResolution(
  expressionMatrix,
  residualMatrix = NULL,
  numberOfComponents = 3,
  initAlgorithm = "simplisma",
  maxIterations = 2000,
  tolerance = 9e-04,
  randomSeed = 91120,
  computeGeneProfiles = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{expressionMatrix}{The expression matrix organized as (*n* X *m*), *n* is the number of genes and *m* is the number of samples.
You should use the expressionMatrix output of prepareData().}

\item{residualMatrix}{The residual matrix, whose dimensions are (*n* X *m*). You should use the residualMatrix output of prepareData().
If the residual matrix is specified, the function uses the error weighted multivariate curve resolution algorithm.
If the residual matrix is NULL, the function uses the non-weighted multivariate curve resolution algorithm.}

\item{numberOfComponents}{The number of number of components (profiles) to be extracted.}

\item{initAlgorithm}{Which algorithm to use to initialize the P/C values. The choice is between "simplisma" (recommended) and "random".}

\item{maxIterations}{The maximum number of iteration for the maximum likelihood algorithm.}

\item{tolerance}{The convergence tolerance. Once this is reached, the algorithm stops.}

\item{randomSeed}{The random seed number, only used if initAlgorithm is "random".}

\item{computeGeneProfiles}{Whether to compute the gene profiler or not.}

\item{verbose}{Whether to display the information about the computation or not.}
}
\value{
The output of is a list containing: Popt, the final estimate of the profile (spectral) matrix, where
vectors (rows) are constrained to be non-negative and unit length. Copt, the final estimate of contribution matrix,
constrained to be non-negative. estimatedMatrix, the final estimated matrix. geneProfiles, the gene profiles (only if computeGeneProfiles = TRUE).
}
\description{
Function to perform multivariate curve resolution by alternating total least squares.
}
\author{
Tobias K. Karakach, Federico Taverna
}
