% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{This function is meant to prepare the input to curve resolution function that calculates the profiles of a
gene expression matrix measured as a function of time. The input data is an expression matrix which must
have a metadata with replicate information. If there are no replicates, the total least squares implementation
of the algorithm cannot be performed.}
\usage{
prepareData(
  expressionData,
  metaData,
  sampleColumn = "Label",
  conditionColumn = "Condition",
  genesAsRowNames = FALSE,
  applyLogTransformation = TRUE
)
}
\arguments{
\item{expressionData}{A data.frame containing the expression data (genes as rows, samples as columns).
The first column of the data.frame must contain the gene identifiers/names.
Alternatively, the gene identifiers/names can be specified as row names.
Duplicated or missing gene identifiers/names will be dropped.}

\item{metaData}{A data.frame containing the metadata (samples as rows, information as columns).
The sample names must match the ones preset in the expression data.}

\item{sampleColumn}{The name of the metadata column that contains the sample names.}

\item{conditionColumn}{The name of the metadata column that contains the condition names.}

\item{genesAsRowNames}{Whether the genes are specified as row names or not.}

\item{applyLogTransformation}{Whether to apply the log2 transformation to the expression data or not.
If the data is already log transformed, the value should be FALSE.}
}
\value{
A list which contains: the processed expression data (expressionMatrix),
the residual matrix (residualMatrix), the error covariance matrix (errorCovarianceMatrix).
}
\description{
This function is meant to prepare the input to curve resolution function that calculates the profiles of a
gene expression matrix measured as a function of time. The input data is an expression matrix which must
have a metadata with replicate information. If there are no replicates, the total least squares implementation
of the algorithm cannot be performed.
}
\author{
Tobias K. Karakach, Federico Taverna
}
