% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplisma.R
\name{simplisma}
\alias{simplisma}
\title{Function to obtain the initial values for P or C.
Ref: W. Windig, J. Guilment, Anal. Chem., 63, 1425 (1991)}
\usage{
simplisma(expressionMatrix, numberOfComponents, noiseFactor = 0.009)
}
\arguments{
\item{expressionMatrix}{The expression matrix organized as (*n* X *m*), *n* is the number of genes and *m* is the number of samples.
You should use the expressionMatrix output of prepareData().}

\item{numberOfComponents}{The number of number of components (profiles) to be extracted.}

\item{noiseFactor}{The noise factor.}
}
\value{
A list which contains the initial estimates for P (Pinit) and C (Cinit).
}
\description{
Function to obtain the initial values for P or C.
Ref: W. Windig, J. Guilment, Anal. Chem., 63, 1425 (1991)
}
\author{
Tobias K. Karakach, Federico Taverna
}
