% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeastExample-data.R
\docType{data}
\name{yeastExample}
\alias{yeastExample}
\title{Example data (Yeast)}
\format{
A list containing the data (expressionData), metadata (metaData), the sample column (sampleColumn) and the condition column (conditionColumn).
}
\usage{
data(yeastExample)
}
\description{
Example data for the EMOGEA package.
}
\examples{
data(yeastExample)
expressionData <- yeastExample$expressionData
metaData <- yeastExample$metaData
sampleColumn <- yeastExample$sampleColumn
conditionColumn <- yeastExample$conditionColumn
}
\keyword{datasets}
